Refactor (Monorepo): Alter configuration of Monorepo
Add yarn workspaces to lerna so that linking dependencies and their
installation process may spike up significantly and so that there is
no duplication of the same package in the project. Apart from that also
included some scripts in global package.json in order to facilitate
lerna launching the same scripts from various packages across the
project.